requirements.txt
mysql-connector
pandas
python-dotenv
setuptools